+++ CHANGELOG.md
@@ -21,6 +21,22 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.1.4
+
+Released Thursday, 28th May 2020.
+
+### Dependencies
+
+- Bumped to `ts-lib-value-objects` v2.x
+- Bumped to TypeScript v3.9
+
+### Fixes
+
+- DataLocation now implements `Value<string>`
+  - previously, this was done by `Filepath` et al
+  - improves support for Liskov substitution
+- export `NotAURLError` error class
+
 ## v0.1.3
 
 Released Thursday, 28th May 2020.